requirements.txt
click==8.1.8
colorama==0.4.6
fake-useragent==2.1.0
fastapi==0.115.11
packaging==24.2
propcache==0.3.0
pycparser==2.22
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
selenium==4.29.0
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
uvicorn==0.34.0
webdriver-manager==4.0.2
websocket-client==1.8.0
cryptography==44.0.2
